[FLINK-39387][Test] Fix flaky scheduler benchmark tests caused by thread assertion failure in async TDD creation
(cherry picked from commit 106acbced18b0da5fa17181ef6a72738e5ab211c)
